move app from client service add redisson metadata redisson (demo code)

+++ deployment-manager/deployment-manager-server/src/main/java/at/sintrum/fog/deploymentmanager/service/ApplicationManagerServiceImpl.java
@@ -11,12 +11,13 @@ import at.sintrum.fog.metadatamanager.api.dto.DockerImageMetadata;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.scheduling.annotation.Async;
+import org.springframework.scheduling.annotation.AsyncResult;
 import org.springframework.stereotype.Service;
 import org.springframework.util.StringUtils;
-import org.springframework.web.bind.annotation.RequestBody;
 
 import java.util.Collections;
 import java.util.UUID;
+import java.util.concurrent.Future;
 
 /**
  * Created by Michael on 2017-06-29.
@@ -31,8 +32,16 @@ public class ApplicationManagerServiceImpl implements ApplicationManagerService 
     private final EnvironmentInfoService environmentInfoService;
     private final ShutdownApplicationService shutdownApplicationService;
     private final DeploymentService deploymentService;
+    private final MetadataServiceImpl metadataService;
 
-    public ApplicationManagerServiceImpl(DockerService dockerService, ImageMetadataApi imageMetadataApi, ContainerMetadataApi containerMetadataApi, DeploymentManagerClientFactory clientFactory, EnvironmentInfoService environmentInfoService, ShutdownApplicationService shutdownApplicationService, DeploymentService deploymentService) {
+    public ApplicationManagerServiceImpl(DockerService dockerService,
+                                         ImageMetadataApi imageMetadataApi,
+                                         ContainerMetadataApi containerMetadataApi,
+                                         DeploymentManagerClientFactory clientFactory,
+                                         EnvironmentInfoService environmentInfoService,
+                                         ShutdownApplicationService shutdownApplicationService,
+                                         DeploymentService deploymentService,
+                                         MetadataServiceImpl metadataService) {
         this.dockerService = dockerService;
         this.imageMetadataApi = imageMetadataApi;
         this.containerMetadataApi = containerMetadataApi;
@@ -40,35 +49,32 @@ public class ApplicationManagerServiceImpl implements ApplicationManagerService 
         this.environmentInfoService = environmentInfoService;
         this.shutdownApplicationService = shutdownApplicationService;
         this.deploymentService = deploymentService;
+        this.metadataService = metadataService;
     }
 
     private static final Logger LOG = LoggerFactory.getLogger(ApplicationManagerServiceImpl.class);
 
 
-    private DockerImageMetadata createStatefulServiceCheckpoint(@RequestBody ApplicationMoveRequest applicationMoveRequest, DockerImageMetadata imageMetadata) {
-        String tag = "checkpoint_" + UUID.randomUUID().toString();
-
-        CommitContainerResult checkpoint = dockerService.commitContainer(new CommitContainerRequest(applicationMoveRequest.getContainerId(), Collections.singletonList(tag)));
-        dockerService.pushImage(new PushImageRequest(checkpoint.getImage(), tag));
-
-        imageMetadata.setImage(checkpoint.getImage());
-        imageMetadata.setTag(tag);
-        imageMetadata.setId(null);  //create new/no update
-        imageMetadata = imageMetadataApi.store(imageMetadata);
-        return imageMetadata;
-    }
-
     private FogOperationResult moveContainerToRemote(ApplicationMoveRequest applicationMoveRequest, DockerImageMetadata imageMetadata) {
         at.sintrum.fog.deploymentmanager.client.api.ApplicationManager applicationManagerClient = clientFactory.createApplicationManagerClient(applicationMoveRequest.getTargetFog());
+        FogOperationResult fogOperationResult = null;
 
-        FogOperationResult fogOperationResult = applicationManagerClient.requestApplicationStart(new ApplicationStartRequest(imageMetadata.getId()));
+        try {
+            fogOperationResult = applicationManagerClient.requestApplicationStart(new ApplicationStartRequest(imageMetadata.getId()));
+        } catch (Exception ex) {
+            LOG.error("Remote deployment manager call failed", ex);
+        }
 
-        if (fogOperationResult.isSuccessful()) {
+        if (fogOperationResult != null && fogOperationResult.isSuccessful()) {
             containerMetadataApi.delete(applicationMoveRequest.getContainerId());
-            dockerService.removeContainer(applicationMoveRequest.getContainerId());
+            if (!dockerService.removeContainer(applicationMoveRequest.getContainerId())) {
+                LOG.error("Failed to delete moved container. Unnecessary resources!");
+            }
         } else {
             LOG.info("Move container failed. Restart original container");
-            dockerService.startContainer(applicationMoveRequest.getContainerId());
+            if (!dockerService.startContainer(applicationMoveRequest.getContainerId())) {
+                LOG.error("Houston, we have a problem! Failed to restart original container!");
+            }
             return new FogOperationResult(applicationMoveRequest.getContainerId(), false, environmentInfoService.getFogBaseUrl(), "move failed, recovered");
         }
 
@@ -77,7 +83,7 @@ public class ApplicationManagerServiceImpl implements ApplicationManagerService 
 
     @Async
     @Override
-    public FogOperationResult start(ApplicationStartRequest applicationStartRequest) {
+    public Future<FogOperationResult> start(ApplicationStartRequest applicationStartRequest) {
         String metadataId = applicationStartRequest.getMetadataId();
 
         LOG.info("Request application start: " + metadataId);
@@ -86,13 +92,25 @@ public class ApplicationManagerServiceImpl implements ApplicationManagerService 
 
         if (imageMetadata == null) {
             LOG.error("Image metadata missing for: " + metadataId);
-            return new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "Image metadata missing.");
+            return new AsyncResult<>(new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "Image metadata missing."));
         } else {
-            dockerService.pullImage(new PullImageRequest(imageMetadata.getImage(), imageMetadata.getTag()));
+            //   if (metadataService.addAppStartRequest(applicationStartRequest)) {
+
+            if (!applicationStartRequest.isSkipPull()) {
+                if (!dockerService.pullImage(new PullImageRequest(imageMetadata.getImage(), imageMetadata.getTag()))) {
+                    return new AsyncResult<>(new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "Failed to pull image"));
+                }
+            } else {
+                LOG.info("Skipped to pull image");
+            }
 
             CreateContainerRequest createContainerRequest = deploymentService.buildCreateContainerRequest(imageMetadata);
 
             CreateContainerResult container = dockerService.createContainer(createContainerRequest);
+
+            if (container == null) {
+                return new AsyncResult<>(new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "Failed to create container"));
+            }
 
             if (container.getWarnings() != null && container.getWarnings().length > 0) {
                 LOG.warn("Warnings during container creation. ID: " + container.getId() + "\n" + String.join("\n, ", container.getWarnings()));
@@ -102,29 +120,35 @@ public class ApplicationManagerServiceImpl implements ApplicationManagerService 
             containerMetadataApi.store(containerMetadata);
 
             //TODO: logging
-            dockerService.startContainer(container.getId());
+            if (!dockerService.startContainer(container.getId())) {
+                return new AsyncResult<>(new FogOperationResult(containerMetadata.getContainerId(), false, environmentInfoService.getFogBaseUrl(), "Failed to start container"));
+            }
+            //   metadataService.finishStartup(applicationStartRequest);
+            return new AsyncResult<>(new FogOperationResult(container.getId(), true, environmentInfoService.getFogBaseUrl()));
 
-            return new FogOperationResult(container.getId(), true, environmentInfoService.getFogBaseUrl());
+            //}
+            //return new AsyncResult<>(new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "Startup skipped"));
         }
     }
 
     @Async
     @Override
-    public FogOperationResult move(ApplicationMoveRequest applicationMoveRequest) {
+    public Future<FogOperationResult> move(ApplicationMoveRequest applicationMoveRequest) {
         ContainerInfo containerInfo = dockerService.getContainerInfo(applicationMoveRequest.getContainerId());
 
         if (containerInfo == null) {
             LOG.warn("Can't move container. Unknown container '" + applicationMoveRequest.getContainerId() + "'");
-            return new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "unknown container");
+            return new AsyncResult<>(new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "unknown container"));
         } else {
             DockerContainerMetadata containerMetadata = containerMetadataApi.getById(containerInfo.getId());
 
             if (containerMetadata == null) {
                 LOG.error("ContainerMetadata missing for container: " + containerInfo.getId());
-                return new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "missing container metadata");
+                return new AsyncResult<>(new FogOperationResult(null, false, environmentInfoService.getFogBaseUrl(), "missing container metadata"));
             } else {
 
                 try {
+                    //TODO: check if shutdown is a good choice. we might need to send some do your stuff make yourself ready to move request
                     if (StringUtils.isEmpty(applicationMoveRequest.getApplicationUrl())) {
                         LOG.warn("ApplicationURL missing. Can't send shutdown request.");
                     } else {
@@ -134,20 +158,37 @@ public class ApplicationManagerServiceImpl implements ApplicationManagerService 
                 } catch (Exception ex) {
                     LOG.error("Request application shutdown failed with: " + ex);
                 }
-                dockerService.stopContainer(applicationMoveRequest.getContainerId());
+                if (!dockerService.stopContainer(applicationMoveRequest.getContainerId())) {
+                    return new AsyncResult<>(new FogOperationResult(containerInfo.getId(), false, environmentInfoService.getFogBaseUrl(), "Failed to stop container"));
+                }
 
                 DockerImageMetadata imageMetadata = imageMetadataApi.getById(containerMetadata.getImageMetadataId());
 
                 if (imageMetadata == null) {
                     LOG.error("ImageMetadata was null for ID: " + containerMetadata.getImageMetadataId());
-                    return new FogOperationResult(containerMetadata.getContainerId(), false, environmentInfoService.getFogBaseUrl(), "image metadata missing");
+                    return new AsyncResult<>(new FogOperationResult(containerMetadata.getContainerId(), false, environmentInfoService.getFogBaseUrl(), "Image metadata missing"));
                 }
 
                 if (!imageMetadata.isStateless()) {
-                    imageMetadata = createStatefulServiceCheckpoint(applicationMoveRequest, imageMetadata);
+                    String tag = "checkpoint_" + UUID.randomUUID().toString();
+
+                    CommitContainerResult checkpoint = dockerService.commitContainer(new CommitContainerRequest(applicationMoveRequest.getContainerId(), Collections.singletonList(tag)));
+
+                    if (checkpoint == null) {
+                        return new AsyncResult<>(new FogOperationResult(containerInfo.getId(), false, environmentInfoService.getFogBaseUrl(), "Failed to create checkpoint"));
+                    }
+
+                    if (!dockerService.pushImage(new PushImageRequest(checkpoint.getImage(), tag))) {
+                        return new AsyncResult<>(new FogOperationResult(containerInfo.getId(), false, environmentInfoService.getFogBaseUrl(), "Failed to push checkpoint"));
+                    }
+
+                    imageMetadata.setImage(checkpoint.getImage());
+                    imageMetadata.setTag(tag);
+                    imageMetadata.setId(null);  //create new/no update
+                    imageMetadata = imageMetadataApi.store(imageMetadata);
                 }
 
-                return moveContainerToRemote(applicationMoveRequest, imageMetadata);
+                return new AsyncResult<>(moveContainerToRemote(applicationMoveRequest, imageMetadata));
             }
         }
     }
